make a baseSqlHook and apply it to mysql, postgres, sqlite
