+++ airflow/hooks/__init__.py
@@ -4,6 +4,7 @@ abstracting the underlying modules
 '''
 from airflow.utils import import_module_attrs as _import_module_attrs
 from airflow.hooks.base_hook import BaseHook as _BaseHook
+from airflow.hooks.base_hook import BaseHook as _BaseSqlHook
 
 _hooks = {
     'hive_hooks': [
