+++ airflow/hooks/mysql_hook.py
@@ -4,10 +4,10 @@ import logging
 
 import MySQLdb
 
-from airflow.hooks.base_hook import BaseHook
+from airflow.hooks.base_hook import BaseSqlHook
 
 
-class MySqlHook(BaseHook):
+class MySqlHook(BaseSqlHook):
     '''
     Interact with MySQL.
     '''
@@ -27,28 +27,6 @@ class MySqlHook(BaseHook):
             conn.password,
             conn.schema)
         return conn
-
-    def get_records(self, sql):
-        '''
-        Executes the sql and returns a set of records.
-        '''
-        conn = self.get_conn()
-        cur = conn.cursor()
-        cur.execute(sql)
-        rows = cur.fetchall()
-        cur.close()
-        conn.close()
-        return rows
-
-    def get_pandas_df(self, sql):
-        '''
-        Executes the sql and returns a pandas dataframe
-        '''
-        import pandas.io.sql as psql
-        conn = self.get_conn()
-        df = psql.read_sql(sql, con=conn)
-        conn.close()
-        return df
 
     def run(self, sql):
         conn = self.get_conn()
